ye feue

+++ src/dht11/dht11.cpp
@@ -3,10 +3,10 @@
 dht11::dht11(uint8_t p) : _pin(p), _dht11(_pin){
     _startTime = 0;
     _periodMs = 2000;
-    _highTemp = 37.0;
-    _lowTemp = 15.0;
-    _highHum = 80.0;
-    _lowHum = 40.0;
+    _highTemp = 37;
+    _lowTemp = 15;
+    _highHum = 80;
+    _lowHum = 40;
 }
 
 void dht11::begin() {}
@@ -14,7 +14,7 @@ void dht11::begin() {}
 void dht11::process() {
     if(millis() - _startTime >= _periodMs) {
         _startTime = millis();
-        float a = 0.0, b = 0.0;
+        int a = 0, b = 0;
         _dhtReading = _dht11.readTemperatureHumidity(a,b);
         if(_dhtReading == 0) {
             _temperatura = a;
@@ -27,7 +27,7 @@ void dht11::process() {
     }
 }
 
-void dht11::setThresholds(float highTemp, float lowTemp, float highHum, float lowHum) {
+void dht11::setThresholds(uint8_t highTemp, uint8_t lowTemp, uint8_t highHum, uint8_t lowHum) {
     _highTemp = highTemp;
     _lowTemp = lowTemp;
     _highHum = highHum;
@@ -51,10 +51,10 @@ bool dht11::pocaHumedad() const {
     return _humedad <= _lowHum;
 }
 
-float dht11::getTemperatura() const {
+uint8_t dht11::getTemperatura() const {
     return _temperatura;
 }
 
-float dht11::getHumedad() const {
+uint8_t dht11::getHumedad() const {
     return _humedad;
 }
